Add small sloped mode (matching PixVoxelAssets)
The current style produced by PixVoxelAssets' isometric rendering is a
smaller version of the existing sloped renders here, but with slightly
different perspective and much smaller output, about as small as it gets
while still being close to legible. Certain details are not always
rendered well with the small mode, like eyes with mostly dark voxels and
a single light voxel for "eye shine," which sometimes render as all
shine or all dark even in the same frame (for two different eyes). The
way around this is mostly just to exaggerate small details, which works
well for pixel art styles anyway. Eyes can just be 4 or more dark voxels
(at least 2 vertically) with the small style and it looks fine.

The colors changed again and should be closer between the sloped version
and the blocky versions. The zombie voxel model also got a small update.