Complete index validity checks; release 0.0.9
The somewhat-automatically-generated code that handles outlining, using
many single-line if/else statements, now has complete checks for
validity of array indices before reading from or writing to them. This
was present in most of this "ugly code", but not all, and #3 was caused
by my failure to validate all the cases. This is fixed now, and some of
the "big" renders should look more correct now since one of the outer
edges was not being rendered before in the double-size-and-up renders.