WIP changes to address #5
This doesn't fully work as I want it to, but iso now can be rendered at multiple sizes like ortho already could. The problem is, it still uses blocky renders like the smaller sizes, while before it was smooth (and only really could be for ortho, since it rendered 1-voxel-to-1-pixel and there's less to worry about with that for ortho shading). Some other things also need adjusting, like size and placement of images.