Center renders inside whitespace
This should be useful for using the generated files in games, since now all the ortho renders can be put onto a square-tiled map and be where they shouuld on a tile, while the same is true of isometric renders for an isometric grid of tiles. Isometric and othographic renders are still different sizes, slightly.